requirements.txt
customtkinter==5.2.0
cloudscraper==1.2.71
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
pydantic==2.5.2
structlog==23.2.0
